requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
langgraph>=0.1.0
chromadb>=0.5.0
pylint>=3.0.0
radon>=6.0.0
streamlit>=1.35.0
python-dotenv>=1.0.0
pytest>=8.0.0
pyyaml>=6.0.0
